Add text manipulation package
This commit adds a text manipulation package to perform two functions:

* Deduplicate characters from a set of data

Character deduplication takes the form of building up a basic set
implementation containing all the characters from the responses. Go doesn't
have a set data structure built-in, and a Map with boolean values appears to be
among the more popular implementations.

* Remove the integers characters from a set of data
It seemed like the most trivial way to implement this was match against a
non-Integer regex. Probably the fastest too since it's leveraging go built-in
regex support.

Signed-off-by: Moses Mendoza <moses@puppet.com>