Implement a client to grab text from URLs
Incoming requests will be formatted in the manner "url,url..." and the client's
responsibility is to obtain the text at the linked urls and compose into a
single body of text for further processing.

We use the built-in http client in Go to accomplish this. Because we could have
many possible URLs, we support concurrent client downloads of all of the URLs
using go routines, a channel, and a WaitGroup. The channel is used to buffer
the downloaded content as we obtain it, while the WaitGroup blocks until all
clients have downloaded their content.

There are several major flaws in this implementation:
1. Nothing is done to prevent oversize request bodies from consuming all
resources. Would need to place a limit on request body size, and possibly make
this configurable/exposed to the user.

2. We will block indefinitely (or until some built-in client timeout if it
exists) if a download hangs. We don't specify any sort of timeout ourselves.

3. We need more tests.

Signed-off-by: Moses Mendoza <moses@puppet.com>