Add a basic http server
Leverage the built-in Go http server to listen on port 8000 for requests and
hand off processing to our handler package. This package could house future
optimization of the server. Note that this hardcodes several values which
shouldn't be: listen address and port for one. A more robust application would
expose these as configuration values for the user to specify.

Signed-off-by: Moses Mendoza <moses@puppet.com>